enorme

+++ ft_strjoin.c
@@ -1,3 +1,14 @@
+/* ************************************************************************** */
+/*                                                                            */
+/*                                                        :::      ::::::::   */
+/*   ft_strjoin.c                                       :+:      :+:    :+:   */
+/*                                                    +:+ +:+         +:+     */
+/*   By: tbouvera <tbouvera@student.42.fr>          +#+  +:+       +#+        */
+/*                                                +#+#+#+#+#+   +#+           */
+/*   Created: 2022/10/17 11:54:43 by tbouvera          #+#    #+#             */
+/*   Updated: 2022/10/18 10:42:36 by tbouvera         ###   ########.fr       */
+/*                                                                            */
+/* ************************************************************************** */
 
 #include "libft.h"
 
@@ -10,7 +21,7 @@ char	*ft_strjoin(char const *s1, char const *s2)
 
 	index = 0;
 	ptr_index = 0;
-	total_size = ft_strlen(s1) + ft_strlen(s2);
+	total_size = (int)ft_strlen(s1) + (int)ft_strlen(s2);
 	ptr = (char *)malloc (total_size * sizeof(char) + 1);
 	if (!ptr)
 		return (NULL);
