+++ ft_strrchr.c
@@ -1,3 +1,14 @@
+/* ************************************************************************** */
+/*                                                                            */
+/*                                                        :::      ::::::::   */
+/*   ft_strrchr.c                                       :+:      :+:    :+:   */
+/*                                                    +:+ +:+         +:+     */
+/*   By: tbouvera <tbouvera@student.42.fr>          +#+  +:+       +#+        */
+/*                                                +#+#+#+#+#+   +#+           */
+/*   Created: 2022/10/17 11:55:13 by tbouvera          #+#    #+#             */
+/*   Updated: 2022/10/18 10:42:28 by tbouvera         ###   ########.fr       */
+/*                                                                            */
+/* ************************************************************************** */
 
 #include "libft.h"
 
@@ -5,7 +16,7 @@ void	*ft_strrchr(const char *str, int c)
 {
 	int	index;
 
-	index = ft_strlen(str);
+	index = (int)ft_strlen(str);
 	while (str[index] != ((char)c))
 	{
 		if (index == 0)
